services/db_service.py
import mysql.connector
from mysql.connector import Error
import logging
from configs.configs import DB_HOST, DB_NAME, DB_USER, DB_PASSWORD

DB_CONFIG = {
    'host': DB_HOST,
    'database': DB_NAME,
    'user': DB_USER, 
    'password': DB_PASSWORD 
}

logger = logging.getLogger(__name__)

def get_connection():
    try:
        connection = mysql.connector.connect(**DB_CONFIG)
        return connection
    except Error as e:
        logger.error(f"Error connecting to MySQL: {e}")
        return None
    
def check_file_exists(filename):
    """Checks if a source filename already exists in the DB."""
    conn = get_connection()
    if conn:
        cursor = conn.cursor()
        query = "SELECT id FROM user_transfers WHERE source_filename = %s"
        cursor.execute(query, (filename,))
        exists = cursor.fetchone() is not None
        cursor.close()
        conn.close()
        return exists
    return False

def insert_user_transfer(data):
    conn = get_connection()
    if conn:
        cursor = conn.cursor()
        
        query = """
        INSERT IGNORE INTO user_transfers 
        (transaction_ref, sender_name, phone_number, phone_confidence, amount, amount_confidence, transaction_date, ocr_confidence, source_filename, user_id)
        VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
        """
        
        try:
            cursor.execute(query, (
                data.get('Transaction Reference'),
                data.get('Sender/Receiver Name'),
                data.get('Mobile Number'),
                data.get('phone_confidence'), 
                data.get('Amount Transferred'),
                data.get('amount_confidence'),   
                data.get('Transaction Date'),
                data.get('ocr_confidence'),
                data.get('source_filename'),
                data.get('user_id') 
            ))
            conn.commit()
           
            if cursor.rowcount > 0:
                logger.info(f"Inserted: {data.get('source_filename')}")
            else:
                logger.info(f"Skipped Duplicate File/Ref: {data.get('source_filename')}")
                
        except Error as e:
            logger.error(f"Failed to insert user transfer: {e}")
        finally:
            cursor.close()
            conn.close()

def insert_bank_transfer(row):
    """Inserts CSV extracted data, skipping if Ref_ID already exists."""
    conn = get_connection()
    if conn:
        cursor = conn.cursor()
        
        query = """
        INSERT IGNORE INTO bank_transfers 
        (transaction_date, amount, ref_id, phone_number)
        VALUES (%s, %s, %s, %s)
        """
        
        # Handle phone number list/formatting
        phone = None
        if isinstance(row['Phone_number'], list) and row['Phone_number']:
             phone = row['Phone_number'][0]
        elif isinstance(row['Phone_number'], str):
             phone = row['Phone_number']

        try:
            cursor.execute(query, (
                row['Date'],
                row['Credit'],
                row['Ref_ID'],
                phone
            ))
            conn.commit()
            
            if cursor.rowcount > 0:
                logger.info(f"Inserted Bank Ref: {row['Ref_ID']}")
            else:
                logger.info(f"Skipped Duplicate Bank Ref: {row['Ref_ID']}")
                
        except Error as e:
            logger.error(f"Failed to insert bank transfer: {e}")
        finally:
            cursor.close()
            conn.close()

def run_matching_logic():
    """
    Executes the matching logic and populates matched/anomaly tables.
    """
    conn = get_connection()
    if not conn: return

    cursor = conn.cursor()
    
    print("--- Running Matching Logic ---")

    # 2. FIND EXACT MATCHES
    match_query = """
        INSERT INTO matched_transactions 
        (user_transfer_id, bank_transfer_id, matched_amount, amount_confidence, matched_phone, phone_confidence)
        SELECT 
            u.id, 
            b.id, 
            b.amount,
            u.amount_confidence,
            b.phone_number,
            u.phone_confidence
        FROM user_transfers u
        JOIN bank_transfers b 
          ON u.amount = b.amount 
          AND (u.phone_number = b.phone_number OR u.phone_number LIKE CONCAT('%', SUBSTRING(b.phone_number, -8)))
        LEFT JOIN matched_transactions m ON u.id = m.user_transfer_id OR b.id = m.bank_transfer_id
        WHERE m.id IS NULL;
        """
    cursor.execute(match_query)
    matches_found = cursor.rowcount
    conn.commit()
    print(f"Matches found and stored: {matches_found}")

    # 3. IDENTIFY USER ANOMALIES
    anomaly_user_query = """
    INSERT INTO anomaly_user_transfers (user_transfer_id, amount, reason)
    SELECT u.id, u.amount, 'No matching bank record found or Low Confidence'
    FROM user_transfers u
    LEFT JOIN matched_transactions m ON u.id = m.user_transfer_id
    WHERE m.id IS NULL
    AND u.id NOT IN (SELECT user_transfer_id FROM anomaly_user_transfers);
    """
    cursor.execute(anomaly_user_query)
    conn.commit()

    # 4. IDENTIFY BANK ANOMALIES
    anomaly_bank_query = """
    INSERT INTO anomaly_bank_transfers (bank_transfer_id, amount, reason)
    SELECT b.id, b.amount, 'No matching user transfer found'
    FROM bank_transfers b
    LEFT JOIN matched_transactions m ON b.id = m.bank_transfer_id
    WHERE m.id IS NULL
    AND b.id NOT IN (SELECT bank_transfer_id FROM anomaly_bank_transfers);
    """
    cursor.execute(anomaly_bank_query)
    conn.commit()
    
    cursor.close()
    conn.close()
    print("Reconciliation complete.")